requirements.txt
python>=3.6.10
tensorflow-gpu>=2.3.0
tf-slim>=1.1.0
numpy>=1.16.1
